requirements_flask.txt
Flask
mysql-connector-python

Flask-Cors


#to run requirements = pip install -r requirements.txt